--- a/chia/full_node/coin_store.py
+++ b/chia/full_node/coin_store.py
@@ -487,7 +487,7 @@ class CoinStore:
                         int(record.coinbase),
                         record.coin.puzzle_hash,
                         record.coin.parent_coin_info,
-                        bytes(record.coin.amount),
+                        bytes(uint64(record.coin.amount)),
                         record.timestamp,
                     )
                 )
@@ -509,7 +509,7 @@ class CoinStore:
                         int(record.coinbase),
                         record.coin.puzzle_hash.hex(),
                         record.coin.parent_coin_info.hex(),
-                        bytes(record.coin.amount),
+                        bytes(uint64(record.coin.amount)),
                         record.timestamp,
                     )
                 )