--- a/chia/util/streamable.py
+++ b/chia/util/streamable.py
@@ -172,6 +172,8 @@ def function_to_convert_one_item(f_type: Type[Any]) -> ConvertFunctionType:
     elif dataclasses.is_dataclass(f_type):
         # Type is a dataclass, data is a dictionary
         return lambda item: dataclass_from_dict(f_type, item)
+    elif hasattr(f_type, "from_json_dict"):
+        return lambda item: f_type.from_json_dict(item)
     elif issubclass(f_type, bytes):
         # Type is bytes, data is a hex string or bytes
         return lambda item: convert_byte_type(f_type, item)