in the `test_full_sync.py` benchmark using full blocks, most of the wall-clock time is spent waiting for I/O. A material amount of that wait is caused by lack of pipelining commands to the aiosqlite thread. The aiosqlite API makes it difficult to pipeline. It does have some special-case commands to reduce round-trips though.

This change makes two changes (in two commits):

1. instead of submitting `execute()` wait and then submit `fetchall()`, use the `execute_fetchall()` function. This saves one round-trip to the aiosqlite thread.
2. Instead of using `SAVEPOINT` and `ROLLBACK TO` commands to support nested transactions, just use plain transactions at the top level, and any failure rolls back the whole thing. This reduces the number of round-trips to the aiosqlite thread by reducing the number of commands we execute.

The result is that the `test_full_sync.py` benchmark takes about 5% less time for the sync benchmark.

| test | before | after | after / before |
| --- | --- | --- | --- |
| full-blocks | 159.84s | 151.86s | 0.95 |
| empty-blocks | 29.65s | 28.11s | 0.948 |